iso9660: avoid warnings introduced by g++ 12.3.0
Otherwise, compile fails with:

    COMPILE  device_model_registry.o
src/server/iso9660/iso9660.cc: In constructor ‘Iso::Sector::Sector(Block::Connection<>&, long unsigned int, long unsigned int, Genode::Entrypoint&)’:
src/server/iso9660/iso9660.cc:62:69: error: arithmetic between floating-point type ‘float’ and enumeration type ‘Iso::Sector::<unnamed enum>’ is deprecated [-Werror=deprecated-enum-float-conversion]
   62 |                                         blk_nr * ((float)blk_size() / BLOCK_SIZE),
      |                                                   ~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
src/server/iso9660/iso9660.cc:63:68: error: arithmetic between floating-point type ‘float’ and enumeration type ‘Iso::Sector::<unnamed enum>’ is deprecated [-Werror=deprecated-enum-float-conversion]
   63 |                                         count * ((float)blk_size() / BLOCK_SIZE));

Ref genodelabs/genode#4827

+++ src/server/iso9660/iso9660.cc
@@ -59,8 +59,8 @@ class Iso::Sector {
 				_p = Block::Packet_descriptor(
 					block.alloc_packet(blk_size() * count),
 					Block::Packet_descriptor::READ,
-					blk_nr * ((float)blk_size() / BLOCK_SIZE),
-					count * ((float)blk_size() / BLOCK_SIZE));
+					blk_nr * ((float)blk_size() / (float) BLOCK_SIZE),
+					count * ((float)blk_size() / (float) BLOCK_SIZE));
 
 				while (!_source.ready_to_submit())
 					ep.wait_and_dispatch_one_io_signal();
